Revert "More improvements to data splitting"
This reverts commit 82fc2011cea28db091bab2986a6a169500914764.
