--- Village123.Shared/Content/Data/placeCategoryData.json
@@ -1,4 +1,0 @@
-[
-  "Furniture",
-  "Crafting"
-]
