--- Village123.Shared/Content/Data/placeData.json
@@ -1,10 +1,0 @@
-{
-  "SingleBed": {
-    "type": "Bed",
-    "category": "Furniture"
-  },
-  "Anvil": {
-    "type": "Workstation",
-    "category": "Crafting"
-  }
-}
