+++ Village123.Shared/Managers/PlaceManager.cs
@@ -3,7 +3,6 @@ using Microsoft.Xna.Framework.Graphics;
 using Newtonsoft.Json;
 using System.Collections.Generic;
 using System.IO;
-using Village123.Shared.Data;
 using Village123.Shared.Entities;
 
 namespace Village123.Shared.Managers
@@ -70,11 +69,11 @@ namespace Village123.Shared.Managers
       }
     }
 
-    public Place Add(PlaceData.Place data, Point point)
+    public Place Add(string name, Point point)
     {
-      var place = new Place(_gwm.GameModel.Content.Load<Texture2D>($"Places/{data.Name}"), point)
+      var place = new Place(_gwm.GameModel.Content.Load<Texture2D>($"Places/{name}"), point)
       {
-        Id = _gwm.IdManager.PlaceId++,
+        Id = _gwm.IdData.PlaceId++,
       };
 
       Places.Add(place);
