+++ Village123.Shared/Village123.Shared.csproj
@@ -10,13 +10,7 @@
     <PackageReference Include="System.Text.Json" Version="8.0.3" />
   </ItemGroup>
   <ItemGroup>
-    <None Update="Content\Data\placeData.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="Content\Data\placeCategoryData.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="Content\Data\placeTypeData.json">
+    <None Update="Content\Data\placeTypesData.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Update="Content\femaleFirstNames.txt">
